pyproject.toml
[project]
name = "skatl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langsmith>=0.7.26",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "tavily-python>=0.7.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath   = ["src"]
